support int8
