+++ onediff_comfy_nodes/modules/oneflow/infer_compiler_registry/register_comfy/CrossAttntionStateDictPatch.py
@@ -1,5 +1,4 @@
 from typing import Dict
-from .CrossAttentionPatch_v2 import is_crossAttention_patch
 
 def process_state_dict_before_saving(state_dict: Dict, graph=None):
     assert graph is not None
@@ -37,37 +36,8 @@ def process_state_dict_before_saving(state_dict: Dict, graph=None):
     return state_dict
 
 
-def apply_patch_after_loading_state_dict(state_dict, input_kwargs=None, graph=None)->Dict:
-    assert graph is not None
-    assert input_kwargs is not None
-
-    try:
-        pass
-        # attn2_patches = (
-        #     input_kwargs.get("transformer_options", {})
-        #     .get("patches_replace", {})
-        #     .get("attn2")
-        # )
-        # if attn2_patches:
-        #     for graph_key, oneflow_graph in state_dict.items():
-        #         name_idx_dict = oneflow_graph["inputs_original"][1][
-        #             "transformer_options"
-        #         ]["patches_replace"]["attn2-name_idx"]
-        #         for k, attn_module in attn2_patches.items():
-        #             for attn_key, attn_tensor in attn_module.state_dict().items():
-        #                 idx = name_idx_dict.get(f"{graph_key}-{k}-{attn_key}")
-        #                 if idx is not None:
-        #                     graph._state_tensor_tuple[idx].copy_(attn_tensor)
-        #                     attn_module.set_attr(
-        #                         attn_key, graph._state_tensor_tuple[idx]
-        #                     )
-                        
-    except Exception as e:
-        print(f"Warning: Failed to apply patch after loading state dict: {e}")
-    
-
-def has_attn2_patch(input_kwargs)->bool:
+def attn2_patch_sum(input_kwargs)->int:
     attn2 = input_kwargs.get("transformer_options", {}).get("patches_replace", {}).get(
         "attn2", {}
     )
-    return len(attn2) > 0
+    return len(attn2)
